fix: extract workspace listing to satisfy the file-length ratchet
